UINT16: use Integer internally to fit value

+++ io.openems.edge.common/test/io/openems/edge/common/modbusslave/ModbusRecordUint16Test.java
@@ -28,7 +28,7 @@ public class ModbusRecordUint16Test {
 
 	@Test
 	public void testValue() {
-		var sut = new ModbusRecordUint16(0, "foo", (short) 12345);
+		var sut = new ModbusRecordUint16(0, "foo", 12345);
 		assertEquals("ModbusRecordUInt16 [value=12345/0x3039, type=uint16]", sut.toString());
 		assertEquals("\"12345\"", sut.getValueDescription());
 	}
